% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{baselineest}
\alias{baselineest}
\title{The baseline hazards function}
\usage{
baselineest(cause, s_zero, Z_int, coef_est)
}
\arguments{
\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{s_zero}{the S_zero in the paper}

\item{Z_int}{the integration of Z_bar}

\item{coef_est}{the coefficent estimate from the fit}
}
\value{
the baseline hazard function estimate
}
\description{
This gives the estimate of baseline hazards function
}
