% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintest.R
\name{betaconfint}
\alias{betaconfint}
\title{the confidence interval for our beta estimator}
\usage{
betaconfint(coef_est, vcov, alpha)
}
\arguments{
\item{coef_est}{the estimate for beta}

\item{vcov}{the variance covariance matrix for beta estimate}

\item{alpha}{the prespecified level}
}
\value{
a list containing the (1 - alpha) level confidence interval
}
\description{
the confidence interval for our beta estimator
}
