% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{betavarest}
\alias{betavarest}
\title{betavarest function}
\usage{
betavarest(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the variance matrix of finite dimensional parameter part
}
\description{
This gives the variance estiamte for heta
}
