% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{coefest}
\alias{coefest}
\title{The finite dimensional coefficients estimator}
\usage{
coefest(N, omega_inv, score_process)
}
\arguments{
\item{N}{the sample size}

\item{omega_inv}{the inverse of omega matrix in the paper}

\item{score_process}{the score process in the paper}
}
\value{
the coefficent estimate from the fit
}
\description{
This function returns us the estimate the finite dimensional part
}
