% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{Dhatt}
\alias{Dhatt}
\title{Dhatt function}
\usage{
Dhatt(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
D_hat part in the paper
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
