% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{Ehatt}
\alias{Ehatt}
\title{Ehatt function}
\usage{
Ehatt(fit, i)
}
\arguments{
\item{fit}{the fitting object after fitting our model}

\item{i}{the ith round of our data}
}
\value{
an integrated function with speed O(n) by recording each time result
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
