% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{Ghatt}
\alias{Ghatt}
\title{Ghatt function}
\usage{
Ghatt(fit, newobsz)
}
\arguments{
\item{fit}{the fitting object after fitting our model}

\item{newobsz}{the new obtained Z value}
}
\value{
G_hat in the paper, the only part that changes with newobsz
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
