% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{Gint}
\alias{Gint}
\title{Gint function}
\usage{
Gint(survtime, censorsurv)
}
\arguments{
\item{survtime}{the event time}

\item{censorsurv}{the estimate for the censoring distribution}
}
\description{
This is the integration of our censorsurv from somewhere to infinity
}
