% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{GZint}
\alias{GZint}
\title{GZ_int function}
\usage{
GZint(survtime, Z_bar, censorsurv)
}
\arguments{
\item{survtime}{the event time}

\item{Z_bar}{defined in the paper}

\item{censorsurv}{the estimate for the censoring distribution}
}
\description{
This is the integration of our censorsurv from somewhere to infinity
}
