% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{hazardpredvarest}
\alias{hazardpredvarest}
\title{hazardpredvarest function}
\usage{
hazardpredvarest(newobsz, fit = NULL)
}
\arguments{
\item{newobsz}{the new obtained Z value}

\item{fit}{the fitting object after fitting our model}
}
\value{
the variance of hazard function at each time point
}
\description{
This gives the variance estimate for our prediction of the hazard function
}
