% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{leadpart}
\alias{leadpart}
\title{leadpart function}
\usage{
leadpart(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the leading part in our variance estimate
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
