% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintest.R
\name{lowerconfint}
\alias{lowerconfint}
\title{the pointwise lower confidence interval for the survival curve}
\usage{
lowerconfint(hazard_pred, hazardpredvar_est, newobsz, alpha)
}
\arguments{
\item{hazard_pred}{the predicted hazard function}

\item{hazardpredvar_est}{the variance of the estimator of the hazard function}

\item{newobsz}{the new obtained Z value}

\item{alpha}{the prespecified level}
}
\value{
the lower (1 - alpha) level pointwise confidence interval for the hazard function
}
\description{
the pointwise lower confidence interval for the survival curve
}
