% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{omegainv}
\alias{omegainv}
\title{Omega inverse function}
\usage{
omegainv(N, survtime, Z, Z_int, cause, comp = FALSE, censorsurv = NULL,
  G_int = NULL, GZ_int = NULL)
}
\arguments{
\item{N}{the sample size}

\item{survtime}{the event time}

\item{Z}{a variable contains all the regressors}

\item{Z_int}{the integration of Z_bar}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause}

\item{comp}{indicator of whether we are under competing risks setting}

\item{censorsurv}{the estimate for the censoring distribution}

\item{G_int}{the integration of G function}

\item{GZ_int}{the integration of GZ}
}
\value{
the Omega matrix appear in the paper
}
\description{
This is for inverse of omega, which is part of the sandwich estimator
}
