% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{pihatt}
\alias{pihatt}
\title{pihatt function}
\usage{
pihatt(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
pi_hatt part in sigma_three, only appears in competing risks setting, in the paper
}
\description{
This gives the pihatt part in the sigmathree estiamte
}
