% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsriadditive.R
\name{print.tsriadditive}
\alias{print.tsriadditive}
\title{the print function associated with our class}
\usage{
\method{print}{tsriadditive}(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\description{
This function will print our coefficients, the variance covariance matrix of
the coeffieients, and the estimate for the baseline hazard function
}
