% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{psihat}
\alias{psihat}
\title{psihat function}
\usage{
psihat(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
psi_hat in the paper
}
\description{
This gives the psihat part in the sigmatwo estiamte
}
