% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{qhatt}
\alias{qhatt}
\title{qhatt function}
\usage{
qhatt(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
q_hatt part in sigma_three, only appears in competing risks setting, in the paper
}
\description{
This gives the qhatt part in the sigmathree estiamte
}
