% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{qprimehatt}
\alias{qprimehatt}
\title{qprimehatt function}
\usage{
qprimehatt(fit, i)
}
\arguments{
\item{fit}{the fitting object after fitting our model}

\item{i}{the i-th round}
}
\value{
update the fit, add qprime_hatt
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
