% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{qprimepihatt}
\alias{qprimepihatt}
\title{qprimepihatt function}
\usage{
qprimepihatt(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the integration of squares of q_t(u) over pi(u) in the paper, introduced by competing risks
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
