% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regadditivefit.R
\name{regadditivefit}
\alias{regadditivefit}
\title{regadditivefit function}
\usage{
regadditivefit(survtime, cause, comp = FALSE, treatment = NULL,
  covariates = NULL)
}
\arguments{
\item{survtime}{the event time}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{comp}{the indicator of whether modeling subdistribution hazard}

\item{treatment}{the treatment variable, can be null}

\item{covariates}{all the observed confounders}
}
\value{
the fitting result, a list containing the cofficients, the baseline function, the variance covariance
function of the coefficients and the byproduct including some pieces during the computing process
}
\description{
fit an additive hazard without using IV method
}
