% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{scoreprocess}
\alias{scoreprocess}
\title{A scoreprocess function}
\usage{
scoreprocess(Z, Z_bar, cause)
}
\arguments{
\item{Z}{a variable contains all the regressors}

\item{Z_bar}{defined in the paper}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}
}
\value{
the score_process
}
\description{
This gives us the scoreprocess function defined in the paper
}
