% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{sigmaone}
\alias{sigmaone}
\title{sigmaone function}
\usage{
sigmaone(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the sigma_one in the paper
}
\description{
This gives the sigmaone part in the variance estiamte
}
