% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{sigmathree}
\alias{sigmathree}
\title{sigmathree function}
\usage{
sigmathree(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the sigma_three in the paper
}
\description{
This gives the sigmathree part in the variance estiamte
}
