% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{sigmatwo}
\alias{sigmatwo}
\title{sigmatwo function}
\usage{
sigmatwo(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
the sigma_two in the paper
}
\description{
This gives the sigmatwo part in the variance estiamte
}
