% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{sone}
\alias{sone}
\title{sone function}
\usage{
sone(survtime, cause = NULL, Z, comp = FALSE, censorsurv = NULL)
}
\arguments{
\item{survtime}{the event time}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{Z}{a variable contains all the regressors}

\item{comp}{the indicator of whether modeling subdistribution hazard}

\item{censorsurv}{the estimate for the censoring distribution}
}
\value{
s_one defined in the paper
}
\description{
The s_one function defined in the paper
}
