% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{survivalpred}
\alias{survivalpred}
\title{A survival prediction function}
\usage{
survivalpred(fit, newtreatment = NULL, newIV = NULL,
  newcovariates = NULL)
}
\arguments{
\item{fit}{the fitting object after fitting our model}

\item{newtreatment}{the new treatment value}

\item{newIV}{new instrumental variable value}

\item{newcovariates}{new observed covariates}
}
\value{
an object recording the corresponding predicted survival curve and corresponding pointwise
confidence interval
}
\description{
The predict function associated with our class
}
