% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{szero}
\alias{szero}
\title{szero function}
\usage{
szero(survtime, cause = NULL, comp = FALSE, censorsurv = NULL)
}
\arguments{
\item{survtime}{the event time}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{comp}{the indicator of whether modeling subdistribution hazard}

\item{censorsurv}{the estimate for the censoring distribution}
}
\value{
s_zero defined in the paper
}
\description{
The s_zero function defined in the paper
}
