% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{szeroint}
\alias{szeroint}
\title{szeroint function}
\usage{
szeroint(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
szero_int part in the paper
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
