% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsriadditivefit.R
\name{tsriadditivefit}
\alias{tsriadditivefit}
\title{fit an additive hazard using two stage residual inclusion method}
\usage{
tsriadditivefit(survtime, cause, comp = FALSE, treatment = NULL,
  IV = NULL, covariates = NULL)
}
\arguments{
\item{survtime}{the event time}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{comp}{the indicator of whether modeling subdistribution hazard}

\item{treatment}{the treatment variable, can be null}

\item{IV}{the instrumental variable}

\item{covariates}{all the observed confounders}
}
\value{
the fitting result, a list containing the cofficients, the baseline function, the variance covariance
}
\description{
fit an additive hazard using two stage residual inclusion method
}
