% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsriadditivefit.R
\name{tsrisurvadditivefit}
\alias{tsrisurvadditivefit}
\title{fit an additive hazard using IV method under survival settings}
\usage{
tsrisurvadditivefit(survtime, cause = NULL, Z)
}
\arguments{
\item{survtime}{the event time}

\item{cause}{the indicator records the cause. Default to all one. Zero means right censoring. Greater than
or equal to two means other cause.}

\item{Z}{a variable contains all the regressors}
}
\value{
the fitting result, a list containing the cofficients, the baseline function and 
the byproduct including some pieces during the computing process
}
\description{
fit an additive hazard using IV method under survival settings
}
