% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varest.R
\name{Yint}
\alias{Yint}
\title{Yint function}
\usage{
Yint(fit)
}
\arguments{
\item{fit}{the fitting object after fitting our model}
}
\value{
Y_int part in the paper
}
\description{
This prepares for the variance estiamte of baseline hazards function
}
