% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{Zbar}
\alias{Zbar}
\title{Zbar function}
\usage{
Zbar(s_zero, s_one)
}
\arguments{
\item{s_zero}{s_zero function}

\item{s_one}{s_one function}
}
\value{
the Z_bar value in the paper, no difference for two settings
}
\description{
A function for computing Z_bar in the paper
}
