% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliaryfunction.R
\name{Zint}
\alias{Zint}
\title{Zint function}
\usage{
Zint(survtime, Z_bar)
}
\arguments{
\item{survtime}{the event time}

\item{Z_bar}{the Z_bar value}
}
\value{
the integration of Z_bar from zero to all the event time t
}
\description{
A function for computing the integration for Z_bar from zero to some time t
}
