% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsriadditive.R
\name{plot.tsriadditive}
\alias{plot.tsriadditive}
\title{plot function associated with our class}
\usage{
\method{plot}{tsriadditive}(fit, newtreatment = NULL, newIV = NULL,
  newcovariates = NULL, alpha = 0.05, unit = "", ...)
}
\arguments{
\item{fit}{the fitting object after fitting our model}

\item{newtreatment}{the new treatment value}

\item{newIV}{new instrumental variable value}

\item{newcovariates}{new observed covariates}

\item{alpha}{the level for confidence interval}

\item{unit}{the time unit we focus}

\item{...}{the other arguments you want to put in the built-in plot function}
}
\description{
this function will plot the predicted curve and corresponding pointwise confidence
interval at level alpha
}
